refactor: useMyPropfileQuery->useMyProfileQuery로 변경
