+++ src/hooks/query/users/useMyprofileEditMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useRecoilValue } from 'recoil';
-import { profileImageUrlState } from '@/hooks/query/users/useMyPropfileQuery';
+import { profileImageUrlState } from '@/hooks/query/users/useMyProfileQuery';
 import usersApi from '@/api/users.api';
 
 // 프로필 수정 => 이미지, 닉네임
